Feat: Create and Develop Payment API using VNPay API

+++ frontend/frontend/src/main/java/com/example/SelfOrderingRestaurant/Dto/Response/DishResponseDTO/GetAllDishesResponseDTO.java
@@ -13,6 +13,6 @@ import java.math.BigDecimal;
 public class GetAllDishesResponseDTO {
     private Integer dishId;
     private String dishName;
-    private BigDecimal price;
+    private Long price;
     private DishStatus status;
 }
